[TASK] Move the XCLASS tests to unit tests
Fixes #187
