+++ Tests/Unit/Domain/Model/Fixtures/XclassFrontendUserGroup.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\FeUserExtraFields\Tests\Functional\Domain\Model\Fixtures;
+namespace OliverKlee\FeUserExtraFields\Tests\Unit\Domain\Model\Fixtures;
 
 use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUserGroup;
